function-calling: split large file functionCalling.ts
